Moved spec execution code into engine
