+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io.kotest.engine/test/TestCaseExecutor.kt
@@ -2,7 +2,6 @@ package io.kotest.engine.test
 
 import io.kotest.core.extensions.TestCaseExtension
 import io.kotest.core.extensions.resolvedTestCaseExtensions
-import io.kotest.core.internal.isEnabled
 import io.kotest.core.spec.invokeAfterInvocation
 import io.kotest.core.spec.invokeAllAfterTestCallbacks
 import io.kotest.core.spec.invokeAllBeforeTestCallbacks
