+++ kotest-framework/kotest-framework-engine/src/commonMain/kotlin/io.kotest.engine/test/enabled.kt
@@ -1,4 +1,4 @@
-package io.kotest.core.internal
+package io.kotest.engine.test
 
 import io.kotest.common.ExperimentalKotest
 import io.kotest.core.config.configuration
@@ -6,6 +6,7 @@ import io.kotest.core.extensions.EnabledExtension
 import io.kotest.core.extensions.resolvedExtensions
 import io.kotest.core.filter.TestFilter
 import io.kotest.core.filter.TestFilterResult
+import io.kotest.core.internal.KotestEngineProperties
 import io.kotest.core.internal.tags.activeTags
 import io.kotest.core.internal.tags.allTags
 import io.kotest.core.internal.tags.isActive
