+++ kotest-framework/kotest-framework-engine/src/jvmMain/kotlin/io/kotest/engine/spec/SpecExecutor.kt
@@ -3,7 +3,7 @@ package io.kotest.engine.spec
 import io.kotest.core.config.configuration
 import io.kotest.core.extensions.SpecExtension
 import io.kotest.core.extensions.resolvedSpecExtensions
-import io.kotest.core.internal.isEnabled
+import io.kotest.engine.test.isEnabled
 import io.kotest.core.internal.resolvedThreads
 import io.kotest.core.spec.IsolationMode
 import io.kotest.core.spec.Spec
