+++ kotest-framework/kotest-framework-engine/src/jvmTest/kotlin/com/sksamuel/kotest/engine/tags/TagsAnnotationInheritenceTest.kt
@@ -4,7 +4,7 @@ import io.kotest.core.Tag
 import io.kotest.core.Tags
 import io.kotest.core.config.configuration
 import io.kotest.core.extensions.TagExtension
-import io.kotest.core.internal.isEnabledInternal
+import io.kotest.engine.test.isEnabledInternal
 import io.kotest.core.spec.Isolate
 import io.kotest.engine.spec.materializeAndOrderRootTests
 import io.kotest.core.spec.style.FunSpec
